+++ xopener/xopener.reg
@@ -1,0 +1,14 @@
+Windows Registry Editor Version 5.00
+
+[HKEY_CURRENT_USER\Software\Classes\xopener]
+@="URL:xopener Protocol"
+"URL Protocol"=""
+
+[HKEY_CURRENT_USER\Software\Classes\xopener\shell\open\command]
+@="C:\\xopener\\xopener.exe \"%1\""
+
+[HKEY_CURRENT_USER\Software\Policies\Google\Chrome\URLAllowList]
+"255"="xopener:*"
+
+[HKEY_CURRENT_USER\Software\Policies\Microsoft\Edge\URLAllowList]
+"255"="xopener:*"
